Fix findbugs violations in netconf
- Field is a mutable collection which should be package protected
- Unchecked/unconfirmed cast of return value from method
- May expose internal representation by returning reference to mutable object
- May expose internal representation by incorporating reference to mutable object
- Field should be package protected
- Parameter must be non-null but is marked as nullable
- Nullcheck of value previously dereferenced
- Inconsistent synchronization
- Method ignores return value
- Boxing/unboxing to parse a primitive
- Boxed value is unboxed and then immediately reboxed
- Unchecked/unconfirmed cast
- Class implements same interface as superclass
- Switch statement found where one case falls through to the next case
- Field not initialized in constructor but dereferenced without null check
- Reliance on default encoding
- Private method is never called
- Method invokes toString() method on a String
- Possible null pointer dereference
- Consider using Locale parameterized version of invoked method
- Field not guarded against concurrent access
- Method does not release lock on all exception paths
- Inefficient use of keySet iterator instead of entrySet iterator
- Load of known null value
- Redundant nullcheck of value known to be non-null
- Non-transient non-serializable instance field in serializable class
- Class is Externalizable but doesn't define a void constructor
- Class is Serializable, but doesn't define serialVersionUID
- Class defines field that masks a superclass field
- Should be a static inner class
- Repeated conditional tests
- Format string should use %n rather than \n
- Method may fail to close stream
- Unconditional wait
- Wait not in loop
- Method concatenates strings using + in a loop
- Dead store to local variable
- Possible null pointer dereference due to return value of called method
- Useless object created

+++ netconf/tools/netconf-testtool/src/main/java/org/opendaylight/netconf/test/tool/Main.java
@@ -17,6 +17,7 @@ import com.google.common.collect.Lists;
 import com.google.common.io.ByteStreams;
 import com.google.common.io.CharStreams;
 import com.google.common.io.Files;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
@@ -42,6 +43,7 @@ import org.opendaylight.netconf.test.tool.config.ConfigurationBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+@SuppressFBWarnings("DM_DEFAULT_ENCODING")
 public final class Main {
     private static final Logger LOG = LoggerFactory.getLogger(Main.class);
 
@@ -50,6 +52,7 @@ public final class Main {
     }
 
     @SuppressWarnings("checkstyle:IllegalCatch")
+    @SuppressFBWarnings({"UW_UNCOND_WAIT", "WA_NOT_IN_LOOP"})
     public static void main(final String[] args) {
         final TesttoolParameters params = TesttoolParameters.parseArgs(args, TesttoolParameters.getParser());
         params.validate();
@@ -84,7 +87,7 @@ public final class Main {
                     } else {
                         try {
                             future.get();
-                        } catch (final ExecutionException e) {
+                        } catch (final ExecutionException | InterruptedException e) {
                             LOG.info("{}. thread failed.", threadNum, e);
                         }
                     }
@@ -102,7 +105,7 @@ public final class Main {
                 configGenerator.updateFeatureFile(generated);
                 configGenerator.changeLoadOrder();
             }
-        } catch (final Exception e) {
+        } catch (RuntimeException | InterruptedException e) {
             LOG.error("Unhandled exception", e);
             netconfDeviceSimulator.close();
             System.exit(1);
@@ -150,9 +153,12 @@ public final class Main {
                 Preconditions.checkState(configDir.mkdirs(), "Unable to create directory " + configDir);
             }
 
-            for (final File file : configDir.listFiles(pathname ->
-                    !pathname.isDirectory() && pathname.getName().startsWith(SIM_DEVICE_CFG_PREFIX))) {
-                Preconditions.checkState(file.delete(), "Unable to clean previous generated file %s", file);
+            final File[] files = configDir.listFiles(pathname ->
+                    !pathname.isDirectory() && pathname.getName().startsWith(SIM_DEVICE_CFG_PREFIX));
+            if (files != null) {
+                for (final File file : files) {
+                    Preconditions.checkState(file.delete(), "Unable to clean previous generated file %s", file);
+                }
             }
 
             try (InputStream stream = Main.class.getResourceAsStream(NETCONF_CONNECTOR_XML)) {
@@ -183,7 +189,7 @@ public final class Main {
                         String configContent = String.format(
                             middleBlueprint, name, address, String.valueOf(openDevice), String.valueOf(!useSsh));
                         configContent = String.format(
-                            "%s%s%d%s\n%s\n", configContent, "<connection-timeout-millis>",
+                            "%s%s%d%s%n%s%n", configContent, "<connection-timeout-millis>",
                             generateConfigsTimeout, "</connection-timeout-millis>", "</module>");
 
                         builder.append(configContent);
